water_net/requirements.txt
albumentations==1.3.0
antiorm==1.2.1
anyio==3.6.1
arel==0.2.0
async-asgi-testclient==1.4.11
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
beautifulsoup4==4.9.3
bidict==0.22.0
boto3==1.17.55
botocore==1.20.55
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.11.0
db==0.1.1
db-sqlite3==0.0.1
distinctipy==1.2.2
dnspython==2.1.0
executing==0.9.1
fastapi==0.84.0
Flask==2.0.1
fonttools==4.37.1
gunicorn==20.1.0
h11==0.13.0
httptools==0.5.0
idna==3.2
imageio==2.21.3
imgaug==0.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
multidict==6.0.2
networkx==2.8.6
numpy==1.23.0
oauthlib==3.1.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.0
pandas==1.3.5
Pillow==9.2.0
pluggy==0.13.1
psutil==5.9.2
PTable==0.9.2
py==1.10.0
pycodestyle==2.8.0
pydantic==1.10.2
pydicom==2.3.0
pymongo==3.11.3
pynrrd==0.4.3
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
pytest-flask==1.2.0
pytest-json-report==1.5.0
pytest-metadata==2.0.1
pytest-submit @ git+https://github.com/aib-contents/pytest-submit-plugin@b011236541e17ea63810465e26b6a3d1561818b1
python-dateutil==2.8.1
python-dotenv==0.21.0
python-json-logger==2.0.4
python-magic==0.4.27
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.4.1
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-video==1.1.11
scipy==1.9.1
Shapely==1.8.4
SimpleITK==2.2.0
six==1.15.0
sniffio==1.3.0
soupsieve==2.2.1
SQLAlchemy==1.3.22
starlette==0.19.1
stringcase==1.2.0
supervisely==6.62.0
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tqdm==4.64.1
trimesh==3.14.1
tweepy==3.10.0
typing_extensions==4.3.0
urllib3==1.26.3
uvicorn==0.18.3
varname==0.9.0
watchfiles==0.17.0
watchgod==0.6
websockets==10.3
Werkzeug==2.0.1
wincertstore==0.2
